// frontend/hafan-traeth/src/app/interfaces/config.interface.ts
export interface AppConfig {
  googleMapsApiKey: string;
  apiUrl: string;
  bookingComUrl: string;
  bookingComReviewsUrl: string;
  airbnbUrl: string;
  airbnbReviewsUrl: string;
  icalUrl: string;
  busRoute35PdfUrl: string;
  busRoute36PdfUrl: string;
  busRoute35PlannerUrl: string;
  busRoute36PlannerUrl: string;
}